+++ cli/commands/resources/builds/base.go
@@ -1,0 +1,22 @@
+package builds
+
+import (
+	"github.com/taubyte/tau-cli/cli/common"
+	"github.com/taubyte/tau-cli/cli/common/options"
+	"github.com/taubyte/tau-cli/env"
+	"github.com/taubyte/tau-cli/i18n"
+	"github.com/urfave/cli/v2"
+)
+
+func (link) Base() (*cli.Command, []common.Option) {
+	selected, err := env.GetSelectedProject()
+	if err != nil {
+		selected = "selected"
+	}
+
+	return common.Base(&cli.Command{
+		Name:      "builds",
+		Aliases:   []string{"jobs"},
+		ArgsUsage: i18n.ArgsUsageName,
+	}, options.NameFlagSelectedArg0(selected))
+}
