+++ cli/commands/resources/builds/types.go
@@ -1,0 +1,13 @@
+package builds
+
+import (
+	"github.com/taubyte/tau-cli/cli/common"
+)
+
+type link struct {
+	common.UnimplementedBasic
+}
+
+func New() common.Basic {
+	return link{}
+}
