+++ cli/commands/resources/builds/vars.go
@@ -1,0 +1,3 @@
+package builds
+
+const defaultTimeFilter = "10m"
