+++ table/builds/list.go
@@ -1,4 +1,4 @@
-package jobsTable
+package buildsTable
 
 import (
 	"fmt"
